requirements.txt
biopython==1.69
cycler==0.10.0
matplotlib==2.0.2
numpy==1.12.1
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
scikit-learn==0.18.1
scipy==0.19.0
six==1.10.0
